project: rename ipso crate to ipso-cli
