+++ Cargo.toml
@@ -1,7 +1,7 @@
 [workspace]
 members = [
-    "ipso",
     "ipso-builtins",
+    "ipso-cli",
     "ipso-core",
     "ipso-diagnostic",
     "ipso-eval",
