--- ipso/src/ipso.rs
@@ -1,1 +1,0 @@
-pub mod run;
